[Fix] Sample tab: Beat calculation was broken in classic / alternative tempo mode.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@7467 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/View_smp.cpp
@@ -439,7 +439,7 @@ void CViewSample::SetCurSel(SmpLength nBegin, SmpLength nEnd)
 				} else
 				{
 					sndFile.RecalculateSamplesPerTick();
-					beats *= sndFile.GetSampleRate() / (Util::mul32to64_unsigned(sndFile.m_PlayState.m_nCurrentRowsPerBeat, sndFile.m_PlayState.m_nMusicSpeed) * Util::mul32to64_unsigned(sndFile.m_PlayState.m_nSamplesPerTick, sampleRate));
+					beats *= sndFile.GetSampleRate() / static_cast<double>(Util::mul32to64_unsigned(sndFile.m_PlayState.m_nCurrentRowsPerBeat, sndFile.m_PlayState.m_nMusicSpeed) * Util::mul32to64_unsigned(sndFile.m_PlayState.m_nSamplesPerTick, sampleRate));
 				}
 				s += mpt::String::Print(", %1 beats)", mpt::fmt::flt(beats, 0, 5));
 			}
